Change numeric fields to BigDecimal, json ignore event type, set version to 1.0.0

+++ src/main/kotlin/com/rodwan/coreapi/dtos/CartItemDto.kt
@@ -1,17 +1,19 @@
 package com.rodwan.coreapi.dtos
 
+import java.math.BigDecimal
+
 data class CartItemDto(
     val productId: String,
     val productName: String,
-    val price: Double,
+    val price: BigDecimal,
     val quantity: Int,
-    val subtotal: Double,
+    val subtotal: BigDecimal,
 ) {
     companion object {
         fun from(
             productId: String,
             productName: String,
-            price: Double,
+            price: BigDecimal,
             quantity: Int,
         ): CartItemDto =
             CartItemDto(
@@ -19,7 +21,7 @@ data class CartItemDto(
                 productName = productName,
                 price = price,
                 quantity = quantity,
-                subtotal = price * quantity,
+                subtotal = price * quantity.toBigDecimal(),
             )
     }
 }
