+++ src/main/kotlin/com/rodwan/coreapi/events/CartCheckedOutEvent.kt
@@ -1,10 +1,11 @@
 package com.rodwan.coreapi.events
 
+import java.math.BigDecimal
 import java.time.Instant
 
 data class CartCheckedOutEvent(
     override val aggregateId: String,
-    val totalAmount: Double,
+    val totalAmount: BigDecimal,
     val itemCount: Int,
     val shippingAddress: String,
     val paymentMethod: String,
